more x64 portability (partially by avd)
git-svn-id: http://miranda.googlecode.com/svn/trunk@9224 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/skin/skinicons.cpp
@@ -404,7 +404,7 @@ int LoadSkinIcons(void)
 	int i, j = 0;
 	char iconName[MAX_PATH], moduleName[MAX_PATH];
     TCHAR modulePath[MAX_PATH];
-	int iconNameSuffIndx;
+	size_t iconNameSuffIndx;
 	DBVARIANT dbv;
 
 	//
